refactor(whispering): name the two conditions its Epicenter build activates
ADR-0190 left one role with two spellings: Honeycrisp said `epicenter-host`
for host ownership while Whispering kept saying `tauri` for both that and
native access. That was a transitional apology, and it does not need to be one.

Whispering's Epicenter build now activates both conditions, because both are
true of it, and its seams split along the line the conditions already draw.
Host-owned: the replica, the credential broker, the deployment choice, the
canonical blob bytes, and the asset base, each of which the Bun host owns by
its own file's account. Tauri-owned: recording, clipboard, notification, HTTP,
shortcuts, and the rest, each of which calls a native command.

Nothing this emits changes, and that is checkable rather than hoped: with both
conditions active every seam resolves to the file it resolved to before, and
the rebuilt output still carries the desktop data route and the host blob path
with no SQLite engine. The web build, which activates neither, is untouched.

What changes is what the files say they are. `tsconfig.tauri.json` becomes
`tsconfig.desktop.json` and checks both conditions, so the leaves stay checked
under the resolver the desktop build really uses.
