+++ apps/whispering/src/lib/platform/auth.epicenter-host.ts
@@ -1,5 +1,5 @@
 import { createDesktopBrokerAuth } from '@epicenter/auth/desktop';
-import { desktopAuthBootstrap } from './desktop-auth-bootstrap.tauri';
+import { desktopAuthBootstrap } from './desktop-auth-bootstrap.epicenter-host';
 import type { PlatformAuth } from './types';
 
 /** One window-local transport client over the process-wide Bun authority. */
