+++ apps/whispering/src/lib/platform/instance.epicenter-host.ts
@@ -1,5 +1,5 @@
 import { createDesktopInstanceSetting } from '@epicenter/auth/desktop';
-import { desktopAuthBootstrap } from './desktop-auth-bootstrap.tauri';
+import { desktopAuthBootstrap } from './desktop-auth-bootstrap.epicenter-host';
 
 /** Desktop deployment writes belong to the Bun authority, never localStorage. */
 export const instanceSetting = createDesktopInstanceSetting({
