+++ apps/whispering/tsconfig.desktop.json
@@ -1,11 +1,13 @@
 {
-	// Typechecks the `#platform/*` tauri condition. `tsconfig.json` resolves the
-	// default (browser) impls, so the desktop platform files and the call sites
-	// that wire them are only checked when the resolver is flipped to `tauri`.
-	// `bun run typecheck` runs both configs; use `typecheck:tauri` for this one
-	// in isolation.
+	// Typechecks the build Epicenter serves, which activates both conditions.
+	// `epicenter-host` selects the leaves whose owner is the Bun host (its
+	// replica, credential, deployment choice, blob bytes, and asset base);
+	// `tauri` selects the leaves that call native commands. `tsconfig.json`
+	// resolves the default (browser) impls, so those files and the call sites
+	// that wire them are only checked here. `bun run typecheck` runs both
+	// configs; use `typecheck:desktop` for this one in isolation.
 	"extends": "./tsconfig.json",
 	"compilerOptions": {
-		"customConditions": ["tauri"]
+		"customConditions": ["epicenter-host", "tauri"]
 	}
 }
